refactor NEON impl into separate file
